Fix a bug that doesn't allow a smaller input
If the user clicked the button and chose to generate a grid that's
smaller than the previous one, it would work, but an error would get
thrown if he tried to press the button again. This is because the
'deleteGrid' function used the length of the array
that contains the squares in the 2nd 'for' that iterates through it,
so not every element was removed from the array because its 'length'
value kept shrinking. The error is fixed by saving that initial value
in a local variable.